requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.34.0
rapidfuzz==3.10.0